Camera can go up smoothly
MainCamera now has a CameraMovement script that mimics the camera going
up.
To use just give any object that moves inside the game the "descender"
script.
Then, press DownArrow and everything moves down at a designated speed.
It's easier to understand if you play with it a little, but long story
short you press DownArrow and and you pretend the camera is going up.